+++ src/framework/react/hooks.ts
@@ -6,26 +6,26 @@
 
 import { useEffect, useMemo, useState } from "react";
 import { useStore as useZustandStore } from "zustand";
+import { loadCodeContribution } from "@/core/code-contrib/load";
 import {
   getContributionPointsMemo,
   getContributionsMemo,
   getExtensionContributionsMemo,
 } from "@/core/contrib-point/get";
-import { loadCodeContribution } from "@/core/code-contrib/load";
+import { getExtensionsMemo } from "@/core/extension/get";
+import { frameworkStore, type FrameworkState } from "@/core/store";
 import type {
   CodeContribution,
   ContributionPoint,
   Extension,
 } from "@/core/types";
-import { frameworkStore, type FrameworkState } from "@/core/store";
-import { getExtensionsMemo } from "@/core/extension/get";
 
-// Reactive hook: use<Name>(args)
+// Hook:            use<Name>(args)
 // Snapshot getter: get<Name>(args)
 // Memoized getter: get<Name>Memo(deps)
 //
 // use<Name>(args) implementation pattern:
-// 1. Collect reactive deps, e.g., use hooks & selectors into framework store
+// 1. Collect reactive deps, e.g., use Hooks & selectors into framework store
 // 2. Return result of get<Name>Memo(deps + args) - no need to use useMemo()
 //
 // get<Name>(args) implementation pattern:
@@ -33,8 +33,10 @@ import { getExtensionsMemo } from "@/core/extension/get";
 // 2. Return result of get<Name>Memo(deps + args)
 
 /**
- * A React hook that gets data from the framework's store
+ * A React Hook that gets data from the framework's store
  * using the given `selector` function.
+ *
+ * Internal helper function.
  *
  * @internal
  * @category React Hooks
@@ -46,7 +48,10 @@ export function useStore<U>(selector: (state: FrameworkState) => U): U {
 }
 
 /**
- * A React hook that provides all registered extensions.
+ * A React Hook that provides all registered extensions.
+ *
+ * See also the {@link core.getExtensions} function,
+ * which is the non-reactive equivalent of this Hook.
  *
  * @category React Hooks
  * @returns An array comprising all extensions
@@ -57,12 +62,23 @@ export function useExtensions(): Extension[] {
 }
 
 /**
- * A React hook that provides all registered contributions for the given
- * contribution point identifier `contribPointId` and optional `contribKey`.
+ * A React Hook that provides all registered contributions for the given
+ * contribution point identifier `contribPointId` and optional contribution key
+ * `contribKey`. The latter is used for contributions that are objects and
+ * where each object property provides a subset of contributions.
+ * In this case, contributions are collected for the same `contribPointId`
+ * and same `contribKey`.
+ *
+ * Usually you won’t use `useContributions` directly in your components.
+ * Instead, you provide your own custom Hook for your contribution point
+ * identified by `contribPointId`.
+ *
+ * See also the {@link core.getContributions} function,
+ * which is the non-reactive equivalent of this Hook.
  *
  * @category React Hooks
  * @param contribPointId - The contribution point identifier
- * @param contribKey - An optional key
+ * @param contribKey - An optional contribution key
  * @returns An array of all extension contributions.
  *   If a contribution is an array, its flattened items are appended
  *   to the return value.
@@ -76,11 +92,22 @@ export function useContributions<T>(
 }
 
 /**
- * A React hook that provides all registered contributions for the given
- * contribution point identifier `contribPointId` and optional `key`.
+ * A React Hook that provides all registered contributions for the given
+ * contribution point identifier `contribPointId` and optional contribution key
+ * `contribKey`. The latter is used for contributions that are objects and
+ * where each object property provides a subset of contributions.
+ * In this case, contributions are collected for the same `contribPointId`
+ * and same `contribKey`.
+ *
+ * Usually you won’t use `useExtensionContributions` directly in your components.
+ * Instead, you provide your own custom Hook for your contribution point
+ * identified by `contribPointId`.
  *
  * The contributions are returned as a mapping from extension identifier
  * to extension contribution.
+ *
+ * See also the {@link core.getExtensionContributions} function,
+ * which is the non-reactive equivalent of this Hook.
  *
  * @category React Hooks
  * @param contribPointId - The contribution point identifier
@@ -100,7 +127,10 @@ export function useExtensionContributions<T>(
 }
 
 /**
- * A React hook that provides all registered contribution points.
+ * A React Hook that provides all registered contribution points.
+ *
+ * See also the {@link core.getContributionPoints} function,
+ * which is the non-reactive equivalent of this Hook.
  *
  * @category React Hooks
  * @returns An array comprising all contributions points
@@ -111,15 +141,24 @@ export function useContributionPoints(): ContributionPoint[] {
 }
 
 /**
- * A React hook that gets a code contribution.
+ * A React Hook that gets a code contribution.
  * It may activate inactive extensions required for the code contribution
  * as a side effect.
+ *
+ * Usually you won’t use `useLoadCodeContribution` directly in your components.
+ * Instead, you provide your own custom Hook for your contribution point
+ * identified by `contribPointId`.
+ *
+ * See also the {@link loadCodeContribution} function,
+ * which is the non-reactive equivalent of this Hook.
  *
  * **Important:** On the very first render of a component, the function returns
  * `undefined`. TODO: Find out how we can we avoid this behaviour.
  *
  * @category React Hooks
  * @typeParam Value - Type of the loaded code contribution value.
+ * @param contribPointId - The contribution point identifier
+ * @param contribId - The code contribution identifier
  * @returns The current state of the code contribution or `undefined`.
  */
 export function useLoadCodeContribution<Value = unknown>(
@@ -162,9 +201,16 @@ export function useLoadCodeContribution<Value = unknown>(
 }
 
 /**
- * Gets the code contribution registrations
+ * Gets a snapshot of the currently registered code contributions
  * as a read-only map. If there are no contributions for the given point,
  * an empty map is returned.
+ *
+ * Note, the Hook will neither load nor activate any contributing extensions.
+ * Instead, it provides the currently registered code contributions only.
+ *
+ * Usually you won’t use `useCodeContributions` directly in your components.
+ * Instead, you provide your own custom Hook for your contribution point
+ * identified by `contribPointId`.
  *
  * @category React Hooks
  * @typeParam Value - Type of the code contribution values in the map.