+++ CHANGES.md
@@ -1,5 +1,8 @@
 # 0.1.1-alpha.0
 
+* Log levels may also be given as strings. [#20]
+* Import logging classes directly from `util`: 
+  `import { Logger, LogLevel } from "@forman2/extendit/util"`.
 * Fixed and enhanced API doc of `compileWhenClause` function. [#16]
 * Clarified effect of `extensionDependencies` in `README.md`.