+++ CHANGES.md
@@ -1,4 +1,4 @@
-# 0.1.1-alpha.0
+# 0.2.0 (in development)
 
 * Enhanced API docs for React Hooks. [#27]
 * Changed type parameter name from `Data` to `Value`. [#24]
@@ -11,6 +11,6 @@
 * Fixed and enhanced API doc of `compileWhenClause` function. [#16]
 * Clarified effect of `extensionDependencies` in `README.md`.
 
-# 0.1.0 from 18.11.2023
+# 0.1.0 (from 18.11.2023)
 
 Initial version.